+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "inform-job-support" extension will be documented in 
 ## [Unreleased]
 ### Added
 - Support Hover to show IO name for JBI files.
+- Support Hover to show Variable name for JBI files.
 
 ## [1.2.0] - 2023-02-08
 ### Added